[refactor]: useUserExists 훅 파일 분리 (#32)
