+++ src/hooks/useUserExists.ts
@@ -1,0 +1,10 @@
+import { useAuthStore, useUserStore } from 'store/index';
+
+export const useUserExists = (): boolean => {
+  const name = useUserStore((state) => state.name);
+  const accessToken = useAuthStore((state) => state.accessToken);
+
+  if (name && accessToken) return true;
+
+  return false;
+};
