+++ src/store/index.ts
@@ -44,12 +44,3 @@ export const useUserStore = create<UserState>()(
     },
   ),
 );
-
-export const useUserExists = (): boolean => {
-  const name = useUserStore((state) => state.name);
-  const accessToken = useAuthStore((state) => state.accessToken);
-
-  if (name && accessToken) return true;
-
-  return false;
-};
